refactor: creating index file to domain folder
