+++ src/common/domain/dtos/auth/index.ts
@@ -1,0 +1,1 @@
+export * from './make-login.dto';
