+++ src/common/domain/dtos/user/index.ts
@@ -1,0 +1,1 @@
+export * from './sign-up-user.dto';
