+++ src/common/domain/interfaces/auth/index.ts
@@ -1,0 +1,1 @@
+export * from './logged-user.interface';
