+++ src/common/domain/interfaces/auth/logged-user.interface.ts
